♻️  (ProgressButton) add a component for timed button
Add a separate component to be able to reuse it.

The count down logic is included in the button as it is core to the component.
The variable `count`used in the useEffect is necessary to be able to
clear the interval after 100 iterations , since using the state would cause
the suppresion and recreation of a new interval at every state update.
Indded we would have to add counter to the useEffect dependencies.