Add interface for teacher (#99)
* ✨(TeacherMeeting) add teacher interface

* 💄(home) differentiate interfaces background

The different interfaces are to similar. The different color background
differentiate the student and the teacher interfaces.

* ✨(StudentMeeting) custom jitsi toolbar

Add a custom toolbar to the Jitsi StudentMeeting

* 💄(Meeting) rename class based on usage

The class BoxMeeting is used in Box and Teacher interface and should be
renamed for clarity.

* 🌐(en) fix traduction error

* 🐛(StudentMeeting) fix to many prejoin menu

The prejoin menu to choose name and connection options was open at
each join. When the jitsi is minimize or maximize it reopens it which is
unwanted. It should only be open once when starting the meeting.