Improve home UI (#116)
* 💄(Home) update button dimensions

The dimensions of the ReturnToLastMeeting button is not
suitable for student and teacher interface.

* 🐛(BowMeeting) fix jitsi-frame size

The JitsiFrame don't need the surrounding div. The removal of
those divs in other interfaces triggers bug due to missing styling.

* 💄(Home) improve home layout

Improve the layout to cover the whole screen with the background color
and position language selector with grid rather than absolute
positioning, which causes overlap on small screen.