🐛(StudentMeeting) fix whiteboard layout
The way the whiteboard is displayed (using vh for the dimensions and
calculating width and haight based on aspect ratio) leads to several
problem. First the image resizes in strange way when the window is
resize but mostly the buttons are overflowing when displaying two images
duing crops.

I had to change the way images are displayed allowing them
to expand up to the container limit. I also refactored some of the state
to handle the circles creation in the ImageViewer as it was making it
more understandable and was removing one state.